+++ src/components/support/BackdropSupport.vue
@@ -48,7 +48,7 @@ const props = withDefaults(defineProps<BackdropSupportProps>(), {
     @apply inset-0 backdrop-blur-none pointer-events-none select-none;
     @apply duration-300 ease-in-out;
     @apply bg-transparent;
-    z-index: 99;
+    z-index: 999;
     transition-property: background-color, backdrop-filter;
 
     &.backdrop-fixed {