+++ src/utils/debounce.ts
@@ -14,7 +14,7 @@ function debounce(callback: Function, ms: number): Function {
     let timeout: number | undefined = void 0;
     return function (this: any, ...args: any[]) {
         window.clearTimeout(timeout);
-        timeout = window.setTimeout(callback.apply(this, args), ms);
+        timeout = window.setTimeout(() => callback.apply(this, args), ms);
     };
 }